+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Add support for set statements.
 - Add graph management, use, create and drop graph.
+- Add support for detach delete.
 
 ## [0.6.1] - 2025-10-28